Move main logic to lib file
